Time: 25 ms (79.19%) | Memory: 8 MB (89.26%) - LeetSync

+++ 1800-concatenation-of-consecutive-binary-numbers/concatenation-of-consecutive-binary-numbers.cpp
@@ -4,7 +4,7 @@ class Solution {
 public:
     int concatenatedBinary(int n) {
         long int ans = 0;
-        long int res = 0;
+        int res = 0;
         for(int i=1;i<n+1;i++)
         {
             if ((i&(i-1)) == 0){res++;}
